Ajout utils : - concatenateStarqueriesFile : permet de concatener des fichiers de starqueries en un seul - concatenateAndCreateSubset : concatene tous les starqueries et fait des subset
Ajout test :
- Bench pour GiantTable/Hexastore/Hexastore no stats avec :
 - LittleSubset + 500k
 - AverageSubset + 500k
 - BigSubset + 500k
 - Degree1 + 500k
 - Degree3-4 + 500k

+++ src/test/java/qengine/storage/RDFHexaStoreVsNoStatisticTest.java
@@ -189,8 +189,8 @@ public class RDFHexaStoreVsNoStatisticTest {
             }
 
             // moyenne par requête dans le batch
-            globalWelfordHexa.add(totalHexa / count);
-            globalWelfordHexaNoStat.add(totalHexaNoStat / count);
+            globalWelfordHexa.add((double) totalHexa / count);
+            globalWelfordHexaNoStat.add((double) totalHexaNoStat / count);
 
             altern = !altern;
         }
